Docs, moved workLog timeEdition, new workLog.
- Basing commands on libcmd, keeping model in taxus.
- New workLog model from scratch.
- Minor changes on the Taxus model. (not managing schema version yet)
- Integrating some output formatter stuff from taxus.
- TODO: Initial workLog commands to create project, ticket, and worklog entry.
- TODO: radical is a long way with parsing, but needs refactoring, and manage storage.
- Various documentation updated.
- Implementing some more handling at libcmd level and session/query at taxus
  level.
- XXX: Removed confparse Values->dict conversion.

+++ radical_numeric_index.py
@@ -38,7 +38,7 @@ Base = declarative_base()
 
 class Comment(Base):
     __tablename__ = 'document_embedded_comments_tinyid'
-    numid = Column(Integer(11), primary_key=True)
+    numid = Column(Integer, primary_key=True)
     comment = Column(Text, index=True)
     comment_hash = Column(String(32), index=True, unique=True)
